// src/service/address-service.ts
import { Address, User } from "@prisma/client";
import {
  AddressResponse,
  CreateAddressRequest,
  GetAddressRequest,
  RemoveAddressRequest,
  toAddressResponse,
  UpdateAddressRequest,
} from "../model/address-model";
import { Validation } from "../validation/validation";
import { AddressValidation } from "../validation/address-validation";
import { ContactService } from "./contact-service";
import { prismaClient } from "../application/database";
import { ResponseError } from "../error/response-error";

export class AddressService {
  static async checkAddress(contactId: number, addressId: number): Promise<Address> {
    const address = await prismaClient.address.findFirst({
      where: {
        id: addressId,
        contact_id: contactId,
      },
    });

    if (!address) {
      throw new ResponseError(404, "address not found");
    }

    return address;
  }

  static async create(user: User, request: CreateAddressRequest): Promise<AddressResponse> {
    const createRequest = Validation.validate(AddressValidation.CREATE, request);

    await ContactService.checkContact(user.username, createRequest.contact_id);

    const address = await prismaClient.address.create({
      data: createRequest,
    });

    return toAddressResponse(address);
  }

  static async get(user: User, request: GetAddressRequest): Promise<AddressResponse> {
    const getRequest = Validation.validate(AddressValidation.GET, request);
    await ContactService.checkContact(user.username, getRequest.contact_id);
    const address = await this.checkAddress(getRequest.contact_id, getRequest.address_id);

    return toAddressResponse(address);
  }

  static async update(user: User, request: UpdateAddressRequest): Promise<AddressResponse> {
    const updateRequest = Validation.validate(AddressValidation.UPDATE, request);

    await ContactService.checkContact(user.username, updateRequest.contact_id);
    await this.checkAddress(updateRequest.contact_id, updateRequest.address_id);

    const { contact_id, address_id, ...addressData } = updateRequest;
    const address = await prismaClient.address.update({
      where: {
        id: address_id,
        contact_id: contact_id,
      },
      data: {
        ...addressData,
      },
    });

    return toAddressResponse(address);
  }

  static async remove(user: User, request: RemoveAddressRequest): Promise<AddressResponse> {
    const removeRequest = Validation.validate(AddressValidation.REMOVE, request);
    await ContactService.checkContact(user.username, removeRequest.contact_id);
    await this.checkAddress(removeRequest.contact_id, removeRequest.address_id);

    const address = await prismaClient.address.delete({
      where: {
        id: removeRequest.address_id,
        contact_id: removeRequest.contact_id,
      },
    });
    return toAddressResponse(address);
  }

  static async list(user: User, contact_id: number): Promise<Array<AddressResponse>> {
    await ContactService.checkContact(user.username, contact_id);

    const addresses = await prismaClient.address.findMany({
      where: {
        contact_id: contact_id,
      },
    });

    return addresses.map((address) => toAddressResponse(address));
  }
}
